[CXF-667] Patch from tomeks applied.  Thanks!
git-svn-id: https://svn.apache.org/repos/asf/incubator/cxf/trunk@540300 13f79535-47bb-0310-9956-ffa450edef68

+++ rt/transports/http/src/main/java/org/apache/cxf/transport/servlet/CXFServlet.java
@@ -225,7 +225,9 @@ public class CXFServlet extends HttpServlet {
     }
 
     public void destroy() {
-        childCtx.destroy();
+        if (childCtx != null) {
+            childCtx.destroy();
+        }
         
         String s = bus.getId();
         BUS_MAP.remove(s);
